[Backend] Moved entitis to models/entities
