+++ backend/src/repositories/user.repository.ts
@@ -1,4 +1,4 @@
 import { appDataSource } from '../config/dataSource';
-import { User } from '../entities/User';
+import { User } from '../models/entities/User';
 
 export const userRepository = appDataSource.getRepository(User).extend({});
